+++ gluecodium/src/test/resources/smoke/errors/output/android-kotlin/com/example/smoke/SomeInternalEnum.kt
@@ -1,0 +1,20 @@
+/*
+
+ *
+ */
+
+@file:JvmName("SomeInternalEnum")
+
+package com.example.smoke
+
+
+internal enum class SomeInternalEnum(@JvmField internal val value: Int) {
+    ONE(1),
+    TWO(2),
+    THREE(3);
+
+    companion object {
+        @JvmField internal val SINGLE = SomeInternalEnum.ONE
+    }
+
+}