+++ gluecodium/src/test/resources/smoke/internal_fields/output/android-kotlin/com/example/smoke/PublicFieldsAllInit.kt
@@ -1,0 +1,27 @@
+/*
+
+ *
+ */
+
+@file:JvmName("PublicFieldsAllInit")
+
+package com.example.smoke
+
+
+class PublicFieldsAllInit {
+    @JvmField var publicField: String
+    @JvmField internal var internalField: String
+
+
+
+    constructor() {
+        this.publicField = "bar"
+        this.internalField = "foo"
+    }
+
+
+
+
+
+}
+