+++ gluecodium/src/test/resources/smoke/listeners/output/android-kotlin/com/example/smoke/InternalListener.kt
@@ -8,7 +8,9 @@
 package com.example.smoke
 
 
-interface InternalListener {
+internal interface InternalListener {
+
+
 
     fun onEvent() : Unit