+++ gluecodium/src/test/resources/smoke/structs/output/android-kotlin/com/example/smoke/Structs.kt
@@ -27,7 +27,7 @@ class Structs : NativeBase {
          */
 
 
-        constructor(phi: Double, r: Double) {
+       constructor(phi: Double, r: Double) {
             val _other = fromPolar(phi, r)
             this.x = _other.x
             this.y = _other.y