+++ gluecodium/src/test/resources/smoke/visibility_platform/output/android-kotlin/com/example/smoke/KotlinInternalClass.kt
@@ -9,7 +9,7 @@ package com.example.smoke
 
 import com.example.NativeBase
 
-class KotlinInternalClass : NativeBase {
+internal class KotlinInternalClass : NativeBase {